+++ frontend/app/settings/page.tsx
@@ -1,0 +1,10 @@
+"use client"
+
+export default function SettingsPage() {
+  return (
+    <div>
+			Settings
+		</div>
+  )
+}
+
